versioning and class added on validator obj
