--- src/feeManager/validator.js
@@ -1,11 +1,0 @@
-const Joi = require('joi');
-const validator = {};
-
-validator.initBeta = async (opts) => {
-  const schema = Joi.object({
-    mnemonic: Joi.string().required(),
-  });
-  await schema.validateAsync(opts);
-};
-
-module.exports = validator;
