+++ src/user/v1/index.js
@@ -1,0 +1,27 @@
+const Request = require('../../request');
+
+class V1 {
+  constructor(auth) {
+    this.auth = auth;
+    this.request = new Request();
+  }
+
+  /**
+   * return current user
+   * @return {object} is response object `{statusCode: 'string', data: 'object'}`
+   */
+  async getUserDetails() {
+    if (!this.auth.authToken) {
+      throw new Error('You must logged In. Try calling auth() method first');
+    }
+    const url = `/user`;
+    const headers = {
+      authToken: this.auth.authToken,
+    };
+    const resp = await this.request.getRequest(url, headers);
+    if (resp instanceof Error) throw resp;
+    return resp.data;
+  }
+}
+
+module.exports = V1;
