+++ src/wallets/v2/validator.js
@@ -1,0 +1,15 @@
+const Joi = require('joi');
+
+class Validator {
+  async balance(opts) {
+    const schema = Joi.object({
+      serviceId: Joi.string().allow('', null),
+      walletID: Joi.string().allow('', null),
+      currency: Joi.string().allow('', null),
+    });
+    await schema.validateAsync(opts);
+  };
+}
+
+
+module.exports = new Validator;
